Revert "alchemy: change file structure to modules for views and models"
