+++ pyramid/scaffolds/alchemy/+package+/__init__.py
@@ -4,8 +4,8 @@ def main(global_config, **settings):
     """ This function returns a Pyramid WSGI application.
     """
     config = Configurator(settings=settings)
-    config.include('pyramid_jinja2')
-    config.include('.models.meta')
+    config.include('pyramid_chameleon')
+    config.include('.models')
     config.add_static_view('static', 'static', cache_max_age=3600)
     config.add_route('home', '/')
     config.scan()
