+++ pyramid/scaffolds/alchemy/+package+/models.py
@@ -1,3 +1,10 @@
+from sqlalchemy import (
+    Column,
+    Index,
+    Integer,
+    Text,
+    )
+
 from sqlalchemy import engine_from_config
 from sqlalchemy.ext.declarative import declarative_base
 from sqlalchemy.orm import sessionmaker
@@ -45,3 +52,11 @@ def get_dbmaker(engine):
     return dbmaker
 
 
+class MyModel(Base):
+    __tablename__ = 'models'
+    id = Column(Integer, primary_key=True)
+    name = Column(Text)
+    value = Column(Integer)
+
+
+Index('my_index', MyModel.name, unique=True, mysql_length=255)
