--- pyramid/scaffolds/alchemy/+package+/models/__init__.py
@@ -1,3 +1,0 @@
-# import all models classes here for sqlalchemy mappers
-# to pick up
-from .mymodel import MyModel
