--- pyramid/scaffolds/alchemy/+package+/models/mymodel.py
@@ -1,16 +1,0 @@
-from .meta import Base
-from sqlalchemy import (
-    Column,
-    Index,
-    Integer,
-    Text,
-)
-
-class MyModel(Base):
-    __tablename__ = 'models'
-    id = Column(Integer, primary_key=True)
-    name = Column(Text)
-    value = Column(Integer)
-
-
-Index('my_index', MyModel.name, unique=True, mysql_length=255)
