+++ pyramid/scaffolds/alchemy/+package+/views.py_tmpl
@@ -3,10 +3,10 @@ from pyramid.view import view_config
 
 from sqlalchemy.exc import DBAPIError
 
-from ..models.mymodel import MyModel
+from .models import MyModel
 
 
-@view_config(route_name='home', renderer='../templates/mytemplate.jinja2')
+@view_config(route_name='home', renderer='templates/mytemplate.pt')
 def my_view(request):
     try:
         one = request.dbsession.query(MyModel).filter(MyModel.name == 'one').first()
